chore: address some clippy lints (#85)
* chore: auto-apply clippy suggestions

* chore: address some clippy lints

* test(context_aware_config::helpers): simplify `test_get_meta_schema` unit test

* test(experimentation_platform): use idiomatic variant names for `Dimensions` enum

* chore(condition_pills): remove unused `impl From<Condition> for String` block

A better way of implementing the same thing would be to implement the `Display`
trait instead. But since the implementation is unused, it has been removed.

* chore(frontend): allow the `clippy::empty_docs` for the entire crate

This lint is raised from functions annotated with the `leptos_macro::component`
macro, which uses doc comments to generate documentation for the component.
For functions without doc comments, the macro seems to generate empty doc
comments, which triggers the lint.

* refactor(service_utils): replace deprecated `base64::decode` usage with `base64::Engine` usage

* chore(experimentation_tests): remove redundant `allow(non_snake_case)` annotation

* chore: address clippy lints for recent upstream commits

* chore: run formatter

* chore(frontend): simplify string equality check

* refactor(cac_client): return the result of the spawned future in `cac_new_client()` FFI function

+++ crates/context_aware_config/src/api/dimension/utils.rs
@@ -22,7 +22,7 @@ pub fn get_all_dimension_schema_map(
                 .compile(&item.schema)
                 .ok()?;
 
-            Some((item.dimension, (compiled_schema, i32::from(item.priority))))
+            Some((item.dimension, (compiled_schema, item.priority)))
         })
         .collect();
